separate comments.js to multiple files
