+++ server/src/routes/comments.js
@@ -1,9 +1,6 @@
 /* eslint-disable camelcase */
 import express from 'express'
-import {
-  getComments, getQuestions, getReactions, getThreadComments,
-  getCommentReactions, addReaction, addQuestion, addComment
-} from '../database'
+import { getComments, getThreadComments, addComment } from '../database'
 import { uuid } from './helpers'
 import { catchErrors } from '../handlers'
 
@@ -13,7 +10,7 @@ const router = express.Router()
 // Retrieve all comments.
 //
 // returns JSON array of all comments in database
-router.get('/comments', catchErrors(async (req, res) => {
+router.get('/', catchErrors(async (req, res) => {
   await getComments().then((rows) => {
     res.send(rows)
   })
@@ -29,7 +26,7 @@ router.get('/comments', catchErrors(async (req, res) => {
 // }
 //
 // Returns 'OK' if comment is succesfully added
-router.post('/comments', catchErrors(async (req, res) => {
+router.post('/', catchErrors(async (req, res) => {
   const { text, user, blob } = req.body
   const id = uuid()
   const threadId = req.body.threadId || uuid()
@@ -43,81 +40,11 @@ router.post('/comments', catchErrors(async (req, res) => {
 // Get comments by threadId
 //
 // returns JSON array of all comments in thread
-router.get('/comments/:threadId', catchErrors(async (req, res) => {
+router.get('/:threadId', catchErrors(async (req, res) => {
   const { threadId } = req.params
   await getThreadComments(threadId).then((rows) => {
     res.send(rows)
   })
-}))
-
-// @api GET /api/questions
-// Retrieve all questions.
-//
-// returns JSON array of all questions in database
-router.get('/questions', catchErrors(async (req, res) => {
-  await getQuestions().then((rows) => {
-    res.send(rows)
-  })
-}))
-
-// @api POST /api/questions
-// adds question to database.
-//
-// Example body @json {
-//   "text": "What?",
-//   "user": "salaattipoika",
-//   "blob": "{\"path\": \"/path/to/element\"}"
-// }
-//
-// Returns 'OK' if question is succesfully added
-router.post('/questions', catchErrors(async (req, res) => {
-  const { text, user, blob } = req.body
-  const id = uuid()
-  const threadId = req.body.threadId || uuid()
-  await addQuestion({
-    id, text, user, threadId, blob,
-  })
-  res.send('OK')
-}))
-
-// @api GET /api/reactions
-// Retrieve all reactions.
-//
-// returns JSON array of all reactions in database
-router.get('/reactions', catchErrors(async (req, res) => {
-  await getReactions().then((rows) => {
-    res.send(rows)
-  })
-}))
-
-// @api GET /api/reactions/:commentId
-// Retrieve all reactions by commentId.
-//
-// returns JSON array of all reactions to comment
-router.get('/reactions/:commentId', catchErrors(async (req, res) => {
-  const { commentId } = req.params
-  await getCommentReactions(commentId).then((rows) => {
-    res.send(rows)
-  })
-}))
-
-// @api POST /api/reactions
-// add reaction to the database.
-//
-// Example body @json {
-//   "emoji": "🍑",
-//   "user": "jaba",
-//   "comment_id": "1bd8052b"
-// }
-//
-// returns 'OK' if reaction is succesfully added
-router.post('/reactions', catchErrors(async (req, res) => {
-  const { emoji, user, commentId } = req.body
-  const id = uuid()
-  await addReaction({
-    id, emoji, user, commentId,
-  })
-  res.send('OK')
 }))
 
 module.exports = router
