+++ server/src/routes/questions.js
@@ -1,0 +1,39 @@
+/* eslint-disable camelcase */
+import express from 'express'
+import { getQuestions, addQuestion } from '../database'
+import { uuid } from './helpers'
+import { catchErrors } from '../handlers'
+
+const router = express.Router()
+
+// @api GET /api/questions
+// Retrieve all questions.
+//
+// returns JSON array of all questions in database
+router.get('/', catchErrors(async (req, res) => {
+  await getQuestions().then((rows) => {
+    res.send(rows)
+  })
+}))
+
+// @api POST /api/questions
+// adds question to database.
+//
+// Example body @json {
+//   "text": "What?",
+//   "user": "salaattipoika",
+//   "blob": "{\"path\": \"/path/to/element\"}"
+// }
+//
+// Returns 'OK' if question is succesfully added
+router.post('/', catchErrors(async (req, res) => {
+  const { text, user, blob } = req.body
+  const id = uuid()
+  const threadId = req.body.threadId || uuid()
+  await addQuestion({
+    id, text, user, threadId, blob,
+  })
+  res.send('OK')
+}))
+
+module.exports = router
