+++ server/src/routes/routes.js
@@ -4,6 +4,8 @@ const router = express.Router()
 
 router.use('/version', require('./version'))
 router.use('/users', require('./users'))
-router.use('/', require('./comments'))
+router.use('/comments', require('./comments'))
+router.use('/questions', require('./questions'))
+router.use('/reactions', require('./reactions'))
 
 module.exports = router
